+++ src/mprMixed.c
@@ -195,7 +195,7 @@ int mncasecmp(MprChar *s1, cchar *s2, ssize n)
         return 1;
     }
     for (rc = 0; n > 0 && *s1 && rc == 0; s1++, s2++, n--) {
-        rc = tolower((int) *s1) - tolower((int) (uchar) *s2);
+        rc = tolower(*s1) - tolower(*s2);
     }
     if (rc) {
         return (rc > 0) ? 1 : -1;
